requirements.txt
pymongo>=3.10.1,<3.11.0
python-dateutil>=2.8.1,<2.9.0
xlsxwriter>=1.2.8,<1.3.0
chardet>=3.0.4,<3.1.0
mutagen>=1.44.0,<1.50.0
flake8>=3.7.9,<3.8.0
coverage>=5.1,<6.0
pylint>=2.5.0,<2.6.0
pytest>=5.4.1,<5.5.0
pathvalidate>=2.3.0,<2.4.0
numpy==1.14.3
scipy>=1.4.1,<1.5.0
psycopg2-binary>=2.8.5,<2.9.0
pandas>=1.0.3,<1.1.0
spotipy>=2.12.0,<2.13.0
fuzzywuzzy>=0.18.0,<0.20.0
python-Levenshtein>=0.12.0,<0.15.0
